requirement.txt
spotipy==2.23.0
pandas==1.5.3
matplotlib==3.7.2
mysql-connector-python==8.0.33